+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Added
 - **General**
     - drf-spectacular support (#2051)
     - ``PROJECTROLES_SUPPORT_CONTACT`` setting support (#2095)
+- **Irodsbackend**
+    - Checksum retrieval in ``get_objects()`` (#2038)
 - **Landingzones**
     - Site read-only mode support (#2051)
     - File type prohibiting by file name suffix (#2064)